Add files via upload
This is the data for the adventure game.  If you want to make your own, you just need to update the data_generator. Do need to keep the categories the same. But you can change riddles, and lot's of other items that are used as prompts for the LLM and diffusers (and now movie generators) to make your world.